requirements.txt
colorama==0.4.6
gradio==4.44.0
openai_whisper==20231117
torch==2.4.1
